+++ go-controller/pkg/ovn/multicast_test.go
@@ -1,0 +1,354 @@
+package ovn
+
+import (
+	"context"
+	"github.com/onsi/ginkgo"
+	"github.com/onsi/gomega"
+	"github.com/onsi/gomega/format"
+
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/config"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/libovsdbops"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/nbdb"
+	addressset "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/address_set"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
+
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/testing/libovsdb"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/types"
+
+	"github.com/urfave/cli/v2"
+
+	v1 "k8s.io/api/core/v1"
+	knet "k8s.io/api/networking/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+type multicastPolicy struct{}
+
+func (p multicastPolicy) getMulticastPolicyExpectedData(ns string, ports []string) []libovsdb.TestData {
+	pg_hash := hashedPortGroup(ns)
+	egressMatch := getACLMatch(pg_hash, getMulticastACLEgrMatch(), lportEgressAfterLB)
+
+	ip4AddressSet, ip6AddressSet := addressset.MakeAddressSetHashNames(ns)
+	mcastMatch := getACLMatchAF(getMulticastACLIgrMatchV4(ip4AddressSet), getMulticastACLIgrMatchV6(ip6AddressSet))
+	ingressMatch := getACLMatch(pg_hash, mcastMatch, lportIngress)
+
+	egressACL := libovsdbops.BuildACL(
+		getMcastACLName(ns, "MulticastAllowEgress"),
+		nbdb.ACLDirectionFromLport,
+		types.DefaultMcastAllowPriority,
+		egressMatch,
+		nbdb.ACLActionAllow,
+		types.OvnACLLoggingMeter,
+		"",
+		false,
+		map[string]string{
+			defaultDenyPolicyTypeACLExtIdKey: string(knet.PolicyTypeEgress),
+		},
+		map[string]string{
+			"apply-after-lb": "true",
+		},
+	)
+	egressACL.UUID = *egressACL.Name + "-UUID"
+
+	ingressACL := libovsdbops.BuildACL(
+		getMcastACLName(ns, "MulticastAllowIngress"),
+		nbdb.ACLDirectionToLport,
+		types.DefaultMcastAllowPriority,
+		ingressMatch,
+		nbdb.ACLActionAllow,
+		types.OvnACLLoggingMeter,
+		"",
+		false,
+		map[string]string{
+			defaultDenyPolicyTypeACLExtIdKey: string(knet.PolicyTypeIngress),
+		},
+		nil,
+	)
+	ingressACL.UUID = *ingressACL.Name + "-UUID"
+
+	lsps := []*nbdb.LogicalSwitchPort{}
+	for _, uuid := range ports {
+		lsps = append(lsps, &nbdb.LogicalSwitchPort{UUID: uuid})
+	}
+
+	pg := libovsdbops.BuildPortGroup(
+		hashedPortGroup(ns),
+		ns,
+		lsps,
+		[]*nbdb.ACL{egressACL, ingressACL},
+	)
+	pg.UUID = pg.Name + "-UUID"
+
+	return []libovsdb.TestData{
+		egressACL,
+		ingressACL,
+		pg,
+	}
+}
+
+var _ = ginkgo.Describe("OVN NetworkPolicy Operations with IP Address Family", func() {
+	const (
+		namespaceName1 = "namespace1"
+		namespaceName2 = "namespace2"
+	)
+	var (
+		app                   *cli.App
+		fakeOvn               *FakeOVN
+		initialDB             libovsdb.TestSetup
+		gomegaFormatMaxLength int
+	)
+
+	ginkgo.BeforeEach(func() {
+		// Restore global default values before each testcase
+		config.PrepareTestConfig()
+		config.IPv4Mode = true
+		config.IPv6Mode = false
+
+		app = cli.NewApp()
+		app.Name = "test"
+		app.Flags = config.Flags
+
+		fakeOvn = NewFakeOVN()
+		gomegaFormatMaxLength = format.MaxLength
+		format.MaxLength = 0
+		initialDB = libovsdb.TestSetup{
+			NBData: []libovsdb.TestData{
+				&nbdb.LogicalSwitch{
+					Name: "node1",
+				},
+			},
+		}
+	})
+
+	ginkgo.AfterEach(func() {
+		fakeOvn.shutdown()
+		format.MaxLength = gomegaFormatMaxLength
+	})
+
+	ginkgo.Context("during execution", func() {
+		for _, m := range getIpModes() {
+			m := m
+			ginkgo.It("tests enabling/disabling multicast in a namespace "+ipModeStr(m), func() {
+				app.Action = func(ctx *cli.Context) error {
+					namespace1 := *newNamespace(namespaceName1)
+
+					fakeOvn.startWithDBSetup(libovsdb.TestSetup{},
+						&v1.NamespaceList{
+							Items: []v1.Namespace{
+								namespace1,
+							},
+						},
+					)
+					setIpMode(m)
+
+					err := fakeOvn.controller.WatchNamespaces()
+					gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					ns, err := fakeOvn.fakeClient.KubeClient.CoreV1().Namespaces().Get(context.TODO(), namespace1.Name, metav1.GetOptions{})
+					gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					gomega.Expect(ns).NotTo(gomega.BeNil())
+
+					// Multicast is denied by default.
+					_, ok := ns.Annotations[util.NsMulticastAnnotation]
+					gomega.Expect(ok).To(gomega.BeFalse())
+
+					// Enable multicast in the namespace.
+					mcastPolicy := multicastPolicy{}
+					expectedData := mcastPolicy.getMulticastPolicyExpectedData(namespace1.Name, nil)
+					ns.Annotations[util.NsMulticastAnnotation] = "true"
+					_, err = fakeOvn.fakeClient.KubeClient.CoreV1().Namespaces().Update(context.TODO(), ns, metav1.UpdateOptions{})
+					gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					gomega.Eventually(fakeOvn.nbClient).Should(libovsdb.HaveData(expectedData...))
+
+					// Disable multicast in the namespace.
+					ns.Annotations[util.NsMulticastAnnotation] = "false"
+					_, err = fakeOvn.fakeClient.KubeClient.CoreV1().Namespaces().Update(context.TODO(), ns, metav1.UpdateOptions{})
+					gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					acls := expectedData[:len(expectedData)-1]
+					gomega.Eventually(fakeOvn.nbClient).Should(libovsdb.HaveData(acls))
+					return nil
+				}
+
+				err := app.Run([]string{app.Name})
+				gomega.Expect(err).NotTo(gomega.HaveOccurred())
+			})
+
+			ginkgo.It("tests enabling multicast in a namespace with a pod "+ipModeStr(m), func() {
+				app.Action = func(ctx *cli.Context) error {
+					namespace1 := *newNamespace(namespaceName1)
+					nPodTestV4 := newTPod(
+						"node1",
+						"10.128.1.0/24",
+						"10.128.1.2",
+						"10.128.1.1",
+						"myPod1",
+						"10.128.1.3",
+						"0a:58:0a:80:01:03",
+						namespace1.Name,
+					)
+					nPodTestV6 := newTPod(
+						"node1",
+						"fd00:10:244::/64",
+						"fd00:10:244::2",
+						"fd00:10:244::1",
+						"myPod2",
+						"fd00:10:244::3",
+						"0a:58:dd:33:05:d8",
+						namespace1.Name,
+					)
+					var tPods []testPod
+					var tPodIPs []string
+					if m.IPv4Mode {
+						tPods = append(tPods, nPodTestV4)
+						tPodIPs = append(tPodIPs, nPodTestV4.podIP)
+					}
+					if m.IPv6Mode {
+						tPods = append(tPods, nPodTestV6)
+						tPodIPs = append(tPodIPs, nPodTestV6.podIP)
+					}
+
+					var pods []v1.Pod
+					for _, tPod := range tPods {
+						pods = append(pods, *newPod(tPod.namespace, tPod.podName, tPod.nodeName, tPod.podIP))
+					}
+
+					fakeOvn.startWithDBSetup(initialDB,
+						&v1.NamespaceList{
+							Items: []v1.Namespace{
+								namespace1,
+							},
+						},
+						&v1.PodList{
+							Items: pods,
+						},
+					)
+					setIpMode(m)
+
+					for _, tPod := range tPods {
+						tPod.populateLogicalSwitchCache(fakeOvn, getLogicalSwitchUUID(fakeOvn.controller.nbClient, "node1"))
+					}
+
+					err := fakeOvn.controller.WatchNamespaces()
+					gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					fakeOvn.controller.WatchPods()
+					ns, err := fakeOvn.fakeClient.KubeClient.CoreV1().Namespaces().Get(context.TODO(), namespace1.Name, metav1.GetOptions{})
+					gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					gomega.Expect(ns).NotTo(gomega.BeNil())
+
+					// Enable multicast in the namespace
+					mcastPolicy := multicastPolicy{}
+					ports := []string{}
+					for _, tPod := range tPods {
+						ports = append(ports, tPod.portUUID)
+					}
+
+					expectedData := mcastPolicy.getMulticastPolicyExpectedData(namespace1.Name, ports)
+					expectedData = append(expectedData, getExpectedDataPodsAndSwitches(tPods, []string{"node1"})...)
+					ns.Annotations[util.NsMulticastAnnotation] = "true"
+					_, err = fakeOvn.fakeClient.KubeClient.CoreV1().Namespaces().Update(context.TODO(), ns, metav1.UpdateOptions{})
+					gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					gomega.Eventually(fakeOvn.nbClient).Should(libovsdb.HaveData(expectedData...))
+					fakeOvn.asf.ExpectAddressSetWithIPs(namespace1.Name, tPodIPs)
+					return nil
+				}
+
+				err := app.Run([]string{app.Name})
+				gomega.Expect(err).NotTo(gomega.HaveOccurred())
+			})
+
+			ginkgo.It("tests adding a pod to a multicast enabled namespace "+ipModeStr(m), func() {
+				app.Action = func(ctx *cli.Context) error {
+					namespace1 := *newNamespace(namespaceName1)
+					nPodTestV4 := newTPod(
+						"node1",
+						"10.128.1.0/24",
+						"10.128.1.2",
+						"10.128.1.1",
+						"myPod1",
+						"10.128.1.3",
+						"0a:58:0a:80:01:03",
+						namespace1.Name,
+					)
+					nPodTestV6 := newTPod(
+						"node1",
+						"fd00:10:244::/64",
+						"fd00:10:244::2",
+						"fd00:10:244::1",
+						"myPod2",
+						"fd00:10:244::3",
+						"0a:58:dd:33:05:d8",
+						namespace1.Name,
+					)
+					var tPods []testPod
+					var tPodIPs []string
+					ports := []string{}
+					if m.IPv4Mode {
+						tPods = append(tPods, nPodTestV4)
+						tPodIPs = append(tPodIPs, nPodTestV4.podIP)
+					}
+					if m.IPv6Mode {
+						tPods = append(tPods, nPodTestV6)
+						tPodIPs = append(tPodIPs, nPodTestV6.podIP)
+					}
+					for _, pod := range tPods {
+						ports = append(ports, pod.portUUID)
+					}
+
+					fakeOvn.startWithDBSetup(initialDB,
+						&v1.NamespaceList{
+							Items: []v1.Namespace{
+								namespace1,
+							},
+						},
+					)
+					setIpMode(m)
+
+					err := fakeOvn.controller.WatchNamespaces()
+					gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					fakeOvn.controller.WatchPods()
+					ns, err := fakeOvn.fakeClient.KubeClient.CoreV1().Namespaces().Get(context.TODO(), namespace1.Name, metav1.GetOptions{})
+					gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					gomega.Expect(ns).NotTo(gomega.BeNil())
+
+					// Enable multicast in the namespace.
+					mcastPolicy := multicastPolicy{}
+					expectedData := mcastPolicy.getMulticastPolicyExpectedData(namespace1.Name, nil)
+					ns.Annotations[util.NsMulticastAnnotation] = "true"
+					_, err = fakeOvn.fakeClient.KubeClient.CoreV1().Namespaces().Update(context.TODO(), ns, metav1.UpdateOptions{})
+					gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					gomega.Eventually(fakeOvn.nbClient).Should(libovsdb.HaveData(append(expectedData, &nbdb.LogicalSwitch{
+						UUID: "node1-UUID",
+						Name: "node1",
+					})...))
+
+					for _, tPod := range tPods {
+						tPod.populateLogicalSwitchCache(fakeOvn, getLogicalSwitchUUID(fakeOvn.controller.nbClient, "node1"))
+						_, err = fakeOvn.fakeClient.KubeClient.CoreV1().Pods(tPod.namespace).Create(context.TODO(), newPod(
+							tPod.namespace, tPod.podName, tPod.nodeName, tPod.podIP), metav1.CreateOptions{})
+						gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					}
+					fakeOvn.asf.EventuallyExpectAddressSetWithIPs(namespace1.Name, tPodIPs)
+					expectedDataWithPods := mcastPolicy.getMulticastPolicyExpectedData(namespace1.Name, ports)
+					expectedDataWithPods = append(expectedDataWithPods, getExpectedDataPodsAndSwitches(tPods, []string{"node1"})...)
+					gomega.Eventually(fakeOvn.nbClient).Should(libovsdb.HaveData(expectedDataWithPods...))
+
+					for _, tPod := range tPods {
+						// Delete the pod from the namespace.
+						err = fakeOvn.fakeClient.KubeClient.CoreV1().Pods(tPod.namespace).Delete(context.TODO(),
+							tPod.podName, *metav1.NewDeleteOptions(0))
+						gomega.Expect(err).NotTo(gomega.HaveOccurred())
+					}
+					fakeOvn.asf.EventuallyExpectEmptyAddressSetExist(namespace1.Name)
+					gomega.Eventually(fakeOvn.nbClient).Should(libovsdb.HaveData(append(expectedData, &nbdb.LogicalSwitch{
+						UUID: "node1-UUID",
+						Name: "node1",
+					})...))
+
+					return nil
+				}
+
+				err := app.Run([]string{app.Name})
+				gomega.Expect(err).NotTo(gomega.HaveOccurred())
+			})
+		}
+	})
+})
